Move deploy.config module out of deploy directory
