--- apiserver/paasng/paasng/engine/deploy/config/__init__.py
@@ -1,12 +1,0 @@
-"""Store configurations related with deployment"""
-from .env_vars import env_vars_providers, get_env_variables
-from .image import ImageCredentialManager
-from .ingress import AppDefaultDomains, AppDefaultSubpaths
-
-__all__ = [
-    'env_vars_providers',
-    'get_env_variables',
-    'ImageCredentialManager',
-    'AppDefaultDomains',
-    'AppDefaultSubpaths',
-]
