+++ apiserver/paasng/paasng/engine/networking/__init__.py
@@ -1,0 +1,4 @@
+"""Manage networking related functions"""
+from .ingress import AppDefaultDomains, AppDefaultSubpaths
+
+__all__ = ["AppDefaultDomains", "AppDefaultSubpaths"]
